+++ pywbemtools/pywbemcli/__init__.py
@@ -27,6 +27,7 @@ from ._cmd_instance import *        # noqa: F403,F401
 from ._cmd_qualifier import *       # noqa: F403,F401
 from ._cmd_server import *          # noqa: F403,F401
 from ._cmd_connection import *      # noqa: F403,F401
+from ._cmd_profile import *         # noqa: F403,F401
 from ._common import *              # noqa: F403,F401
 from ._pywbem_server import *       # noqa: F403,F401
 from ._context_obj import *         # noqa: F403,F401
