+++ pywbemtools/pywbemcli/_cmd_server.py
@@ -26,7 +26,7 @@ from __future__ import absolute_import, print_function
 
 import click
 
-from pywbem import ValueMapping, Error
+from pywbem import Error
 
 from .pywbemcli import cli
 from ._common import format_table, raise_pywbem_error_exception, \
@@ -110,88 +110,6 @@ def server_info(context):
     The information includes CIM namespaces and server brand.
     """
     context.execute_cmd(lambda: cmd_server_info(context))
-
-
-@server_group.command('profiles', cls=PywbemcliCommand,
-                      options_metavar=CMD_OPTS_TXT)
-@click.option('-o', '--organization', type=str, metavar='ORG-NAME',
-              required=False,
-              help='Filter by the defined organization. (ex. -o DMTF')
-@click.option('-p', '--profile', type=str, metavar='PROFILE-NAME',
-              required=False,
-              help='Filter by the profile name. (ex. -p Array')
-@add_options(help_option)
-@click.pass_obj
-def server_profiles(context, **options):
-    """
-    List management profiles advertized by the server.
-
-    Retrieve the CIM instances representing the WBEM management profiles
-    advertized by the WBEM server, and display information about each profile.
-    WBEM management profiles are defined by DMTF and SNIA and define the
-    management functionality that is available.
-
-    The retrieved profiles can be filtered using the --organization and
-    --profile options.
-
-    The output is formatted as a table showing the organization, name, and
-    version for each profile. The --output-format option is ignored unless it
-    specifies a table format.
-    """
-    context.execute_cmd(lambda: cmd_server_profiles(context, options))
-
-
-@server_group.command('centralinsts', cls=PywbemcliCommand,
-                      options_metavar=CMD_OPTS_TXT)
-@click.option('-o', '--organization', type=str, metavar='ORG-NAME',
-              required=False,
-              help='Filter by the defined organization. (ex. -o DMTF')
-@click.option('-p', '--profile', type=str, metavar='PROFILE-NAME',
-              required=False,
-              help='Filter by the profile name. (ex. -p Array')
-@click.option('--cc', '--central-class', 'central_class', type=str,
-              metavar='CLASSNAME', required=False,
-              help='Optional. Required only if profiles supports only '
-              'scopig methodology')
-@click.option('--sc', '--scoping-class', 'scoping_class', type=str,
-              metavar='CLASSNAME', required=False,
-              help='Optional. Required only if profiles supports only '
-              'scopig methodology')
-@click.option('--sp', '--scoping-path', 'scoping_path', type=str,
-              metavar='CLASSLIST', required=False, multiple=True,
-              help='Optional. Required only if profiles supports only '
-              'scopig methodology. Multiples allowed')
-@click.option('--rd', '--reference-direction', 'reference_direction',
-              type=click.Choice(['snia', 'dmtf']),
-              default='dmtf',
-              show_default=True,
-              help='Navigation direction for association.')
-@add_options(help_option)
-@click.pass_obj
-def centralinsts(context, **options):
-    """
-    List central instances of mgmt profiles on the server.
-
-    Retrieve the CIM instances that are central instances of the specified
-    WBEM management profiles, and display these instances. By default, all
-    management profiles advertized on the server are used. The profiles
-    can be filtered by using the --organization and --profile options.
-
-    The central instances are determined using all methodologies defined
-    in DSP1033 V1.1 in the order of GetCentralInstances, central class,
-    and scoping class methodology.
-
-    Profiles that only use the scoping class methodology require the
-    specification of the --central-class, --scoping-class, and --scoping-path
-    options because additional information is needed to perform the scoping
-    class methodology.
-
-    The retrieved central instances are displayed along with the organization,
-    name, and version of the profile they belong to, formatted as a table.
-    The --output-format general option is ignored unless it specifies a table
-    format.
-    """
-    context.execute_cmd(lambda: cmd_server_centralinsts(context, options))
 
 
 ###############################################################
@@ -286,113 +204,5 @@ def cmd_server_info(context):
                                 title='Server General Information',
                                 table_format=output_format))
 
-    except Error as er:
-        raise_pywbem_error_exception(er)
-
-
-def get_profile_info(org_vm, inst):
-    """
-    Get the org, name, and version from the profile instance and
-    return them as a tuple.
-    """
-    org = org_vm.tovalues(inst['RegisteredOrganization'])
-    name = inst['RegisteredName']
-    vers = inst['RegisteredVersion']
-    return org, name, vers
-
-
-def profile_row_sortkey(row):
-    """
-    Return a sort key for profiles that sorts by org name, profile name, and
-    profile version (numeric).
-
-    row is a tuple(org, name, version) where version is a version string such
-    as "1.2.3" or "1.2.3a".
-    """
-    version_parts = []
-    for v_str in row[2].split('.'):
-        try:
-            v_int = int(v_str)
-        except ValueError:
-            v_int = v_str
-        version_parts.append(v_int)
-    return row[0], row[1], tuple(version_parts)
-
-
-def cmd_server_profiles(context, options):
-    """
-    Display general overview of info from current WBEM server
-    """
-    output_format = validate_output_format(context.output_format, 'TABLE')
-
-    server = context.wbem_server
-    try:
-        found_server_profiles = server.get_selected_profiles(
-            registered_org=options['organization'],
-            registered_name=options['profile'])
-
-        org_vm = ValueMapping.for_property(server,
-                                           server.interop_ns,
-                                           'CIM_RegisteredProfile',
-                                           'RegisteredOrganization')
-        rows = []
-        for inst in found_server_profiles:
-            row = get_profile_info(org_vm, inst)
-            rows.append(row)
-        rows.sort(key=profile_row_sortkey)
-        headers = ['Organization', 'Registered Name', 'Version']
-
-        click.echo(format_table(rows, headers,
-                                title='Advertised management profiles:',
-                                table_format=output_format))
-
-    except Error as er:
-        raise_pywbem_error_exception(er)
-
-
-def cmd_server_centralinsts(context, options):
-    """
-    Display general information on the central instances of one or more
-    profiles.
-    """
-    output_format = validate_output_format(context.output_format, ['CIM',
-                                                                   'TABLE'])
-
-    server = context.wbem_server
-    try:
-        found_server_profiles = server.get_selected_profiles(
-            registered_org=options['organization'],
-            registered_name=options['profile'])
-
-        org_vm = ValueMapping.for_property(server,
-                                           server.interop_ns,
-                                           'CIM_RegisteredProfile',
-                                           'RegisteredOrganization')
-        rows = []
-        for inst in found_server_profiles:
-            pi = get_profile_info(org_vm, inst)
-            row = [":".join(pi)]
-            try:
-                ci = server.get_central_instances(
-                    inst.path,
-                    central_class=options['central_class'],
-                    scoping_class=options['scoping_class'],
-                    scoping_path=options['scoping_path'],
-                    reference_direction=options['reference_direction'])
-                row.append("\n".join([str(p) for p in ci]))
-            # mark current inst as failed and continue
-            except Exception as ex:  # pylint: disable=broad-except
-                click.echo('Exception: {} {}'.format(row, ex))
-                row.append("Failed")
-            rows.append(row)
-
-        # sort by org
-        rows.sort(key=lambda x: (x[0]))
-        headers = ['Profile', 'Central Instance paths']
-
-        click.echo(format_table(rows,
-                                headers,
-                                title='Advertised Central Instances:',
-                                table_format=output_format))
     except Error as er:
         raise_pywbem_error_exception(er)
