+++ tests/unit/test_server_cmds.py
@@ -47,29 +47,14 @@ SERVER_HELP_LINES = [
     'Command group for WBEM servers.',
     CMD_OPTION_HELP_HELP_LINE,
     'brand             Get the brand of the server.',
-    'centralinsts      List central instances of mgmt profiles on the server.',
     'info              Get information about the server.',
     'interop           Get the Interop namespace of the server.',
     'namespaces        List the namespaces of the server.',
-    'profiles          List management profiles advertized by the server.',
 ]
 
 SERVER_BRAND_HELP_LINES = [
     'Usage: pywbemcli [GENERAL-OPTIONS] server brand',
     'Get the brand of the server.',
-    CMD_OPTION_HELP_HELP_LINE,
-]
-
-
-SERVER_CENTRAL_INSTS_HELP_LINES = [
-    'Usage: pywbemcli [GENERAL-OPTIONS] server centralinsts  [COMMAND-OPTIONS]',
-    'List central instances of mgmt profiles on the server.',
-    '-o, --organization ORG-NAME Filter by the defined organization',
-    '-p, --profile PROFILE-NAME Filter by the profile name',
-    '--cc, --central-class CLASSNAME Optional. Required only if profiles',
-    '--sc, --scoping-class CLASSNAME Optional. Required only if profiles',
-    '--sp, --scoping-path CLASSLIST Optional. Required only if profiles',
-    '--rd, --reference-direction [snia|dmtf] Navigation direction',
     CMD_OPTION_HELP_HELP_LINE,
 ]
 
@@ -145,18 +130,6 @@ TEST_CASES = [
       'test': 'innows'},
      None, OK],
 
-    ['Verify server command get-centralinsts  --help response',
-     ['centralinsts', '--help'],
-     {'stdout': SERVER_CENTRAL_INSTS_HELP_LINES,
-      'test': 'innows'},
-     None, OK],
-
-    ['Verify server command centralinsts  -h response',
-     ['centralinsts', '-h'],
-     {'stdout': SERVER_CENTRAL_INSTS_HELP_LINES,
-      'test': 'innows'},
-     None, OK],
-
     ['Verify server command info --help response',
      ['info', '--help'],
      {'stdout': SERVER_INFO_HELP_LINES,
@@ -190,18 +163,6 @@ TEST_CASES = [
     ['Verify server command namespaces -h response',
      ['namespaces', '-h'],
      {'stdout': SERVER_NAMESPACES_HELP_LINES,
-      'test': 'innows'},
-     None, OK],
-
-    ['Verify server command profiles --help response',
-     ['profiles', '--help'],
-     {'stdout': SERVER_PROFILES_HELP_LINES,
-      'test': 'innows'},
-     None, OK],
-
-    ['Verify server command profiles -h response',
-     ['profiles', '-h'],
-     {'stdout': SERVER_PROFILES_HELP_LINES,
       'test': 'innows'},
      None, OK],
 
@@ -267,87 +228,6 @@ TEST_CASES = [
       'rc': 1,
       'test': 'innows'},
      MOCK_SERVER_MODEL, OK],
-
-    ['Verify server command profiles',
-     {'args': ['profiles'],
-      'general': ['-d', 'interop', '-o', 'simple']},
-     {'stdout': ['Advertised management profiles:',
-                 'Organization    Registered Name       Version',
-                 '--------------  --------------------  ---------',
-                 'DMTF            Component             1.4.0',
-                 'DMTF            Indications           1.1.0',
-                 'DMTF            Profile Registration  1.0.0',
-                 'SNIA            Array                 1.4.0',
-                 'SNIA            SMI-S                 1.2.0',
-                 'SNIA            Server                1.1.0',
-                 'SNIA            Server                1.2.0',
-                 'SNIA            Software              1.4.0'],
-      'rc': 0,
-      'test': 'innows'},
-     MOCK_SERVER_MODEL, OK],
-
-    ['Verify server command profiles, filtered by org',
-     {'args': ['profiles', '-o', 'DMTF'],
-      'general': ['-d', 'interop', '-o', 'simple']},
-     {'stdout': ['Advertised management profiles:',
-                 'Organization    Registered Name       Version',
-                 '--------------  --------------------  ---------',
-                 'DMTF            Component             1.4.0',
-                 'DMTF            Indications           1.1.0',
-                 'DMTF            Profile Registration  1.0.0'],
-      'rc': 0,
-      'test': 'innows'},
-     MOCK_SERVER_MODEL, OK],
-
-    ['Verify server command profiles, filtered by org, long',
-     {'args': ['profiles', '--organization', 'DMTF'],
-      'general': ['-d', 'interop', '-o', 'simple']},
-     {'stdout': ['Advertised management profiles:',
-                 'Organization    Registered Name       Version',
-                 '--------------  --------------------  ---------',
-                 'DMTF            Component             1.4.0',
-                 'DMTF            Indications           1.1.0',
-                 'DMTF            Profile Registration  1.0.0'],
-      'rc': 0,
-      'test': 'innows'},
-     MOCK_SERVER_MODEL, OK],
-
-    ['Verify server command profiles, filtered by name',
-     {'args': ['profiles', '-p', 'Profile Registration'],
-      'general': ['-d', 'interop', '-o', 'simple']},
-     {'stdout': ['Advertised management profiles:',
-                 'Organization    Registered Name       Version',
-                 '--------------  --------------------  ---------',
-                 'DMTF            Profile Registration  1.0.0'],
-      'rc': 0,
-      'test': 'innows'},
-     MOCK_SERVER_MODEL, OK],
-
-    ['Verify server command profiles, filtered by org, long',
-     {'args': ['profiles', '--profile', 'Profile Registration'],
-      'general': ['-d', 'interop', '-o', 'simple']},
-     {'stdout': ['Advertised management profiles:',
-                 'Organization    Registered Name       Version',
-                 '--------------  --------------------  ---------',
-                 'DMTF            Profile Registration  1.0.0'],
-      'rc': 0,
-      'test': 'innows'},
-     MOCK_SERVER_MODEL, OK],
-
-    ['Verify server command centralinsts based on wbem server mock.',
-     {'args': ['centralinsts', '-o', 'SNIA',
-               '-p', 'Server'],
-      'general': ['-d', 'interop', '-o', 'simple']},
-     {'stdout': ['Advertised Central Instances:',
-                 'Profile',
-                 'Central Instance paths',
-                 'SNIA:Server:1.1.0',
-                 'SNIA:Server:1.2.0',
-                 'interop:MCK_StorageComputerSystem'],
-      'rc': 0,
-      'test': 'innows'},
-     MOCK_SERVER_MODEL, OK],
-
 
     ['Verify server command info with mock server',
      {'args': ['info'],
